BUG #12997. Removed shared lib depdendencies for protoc executable.
If protoc executable depends on other shared libs, then it becomes challenging
to run the executable during build stage of projects such as ParaView since it
requires that the environment is setup correctly for protoc executable to run
(esp. on Windows).

To avoid such issues, we minimize the depedencies for the protoc executable
and bringing in all code statically for the executable as much as possible.